Revert "feat/refactor: add more felte-specific utils, also moved validation's…"
This reverts commit 3ab7b5b712199bfa57e9a440b2927b69ca83726e.
