+++ src/components/form/Form.tsx
@@ -8,10 +8,8 @@ import {
 import { twMerge } from "tailwind-merge";
 import { IComponentBaseProps } from "../types";
 import FormLabel from "./Label";
-import ValidatedForm, { type ValidatedFormProps } from "./ValidatedForm";
-import { useFormValidation } from "./FormValidationContext";
-
-export { type ValidatedFormProps };
+import ValidatedForm, { useFormValidation } from "./ValidatedForm";
+export { type ValidatedFormProps } from "./ValidatedForm";
 
 export type FormProps = Omit<JSX.HTMLAttributes<HTMLFormElement>, "ref"> &
   IComponentBaseProps;
